no more includes

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,7 +1,4 @@
 #include "function_pointers.h"
-#include <stddef.h>
-#include <stdio.h>
-#include <stdlib.h>
 
 /**
  * array_iterator - executes a function on each elemnt of an array
@@ -11,7 +8,7 @@
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	size_t i;
+	unsigned long int i;
 
 	for (i = 0; i < size; i++)
 		(*action)(array[i]);
